+++ tool/check.sh
@@ -21,4 +21,4 @@ echo "    Running tests..."
 flutter test --reporter=compact --no-pub
 
 echo --- fCheck
-dart run ./bin/fcheck.dart --svg .
+dart run ./bin/fcheck.dart --svg --svgfolder .